1.3.5 Fair Management System  Invite to Fair Name: Invite to Fair Participating Actors: Admin, Guidance Counselor Entry Condition: The admin or Guidance Counselor makes a fair invitation. Exit Condition: The invitation form is submitted for evaluation or saved as a draft. Flow of Events: The actor selects the “Invite to Fair” option The system provides a form field with the required details for the invitation such as the invitee’s name, and contact information. The actor fills out and submits the form. The system confirms submission and stores the form in the database.
View Fair Applications
Name: View Fair Applications
Participating Actors: Admin, President
Entry Condition: The admin or President chooses to view all fair applications.
Exit Condition: The list of fair applications is displayed, or the actor chooses to exit.
Flow of Events:
The actor selects the “View Fair Applications” option.
The system retrieves all of the submitted forms from the database.
The system displays all of the applications with the required details such as school names, application statutes, and desired dates.
The actor can review each application by viewing detailed information about the specific application or sort the applications according to various criteria such as school importance and date.
The system updates the display according to the actor’s choices.
The actor completes the review process or exits the page.

View Fairs
Name: View Fairs
Participating Actors: Admin, President
Entry Condition: The Admin or President selects the option to view all accepted fairs.
Exit Condition: The list of accepted fair applications is displayed, or the actor chooses to exit.
Flow of Events:
The actor selects the “View Fairs” option.
The system retrieves all of the accepted fairs from the database.
The system displays all of the accepted fairs with the required details such as school names, number of students, and dates.
The actor can review each fair by viewing detailed information or sort the fairs according to various criteria such as school importance and date.
The system updates the display according to the actor’s choices.
The actor completes the review process or exits the page.

Accept Leading the Fair
Name: Accept Leading the Fair
Participating Actors: Admin, Guide
Entry Condition: The Admin or Guide chooses to review available fair leading opportunities.
Exit Condition: The guide accepts a leading position for a fair, the admin chooses a guide to lead the fair, or the actor exits without making a selection.
Flow of Events:
The actor selects the “Accept Leading the Fair” option.
The system retrieves available fairs needing a guide from the database.
The system displays all of the fairs that need a lead with the required details such as school names, fair statutes, and desired dates.
The actor can review each fair by viewing detailed information or sort the fairs according to various criteria such as school importance and date.
The system records the guide’s acceptance and updates the fair’s status assigned to the guide.
The system confirms successful assignment or displays an error message.